requirements.txt
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
jinja2>=3.1.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.20
passlib[bcrypt]>=1.7.4
fastapi-login>=0.4.1
python-multipart>=0.0.6
